% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_directories.R
\name{generate_directories}
\alias{generate_directories}
\title{Generate directory if it does not exist}
\usage{
generate_directories(dir_path)
}
\arguments{
\item{dir_path}{path to the new directory}
}
\value{

}
\description{
Generate directory if it does not exist
}
\examples{

generate_directories("./new/dir")
}
